Sezione 13: parte 1 - Risolti alcuni problemi di visualizzazione - DistributedCache non più usata - Creazione form di ricerca dati + funzionalità ricerca - Paginare i risultati in elenco - Creazione model binder personalizzato - Mantenere lo stato: CourseListViewModel

+++ src/MyCourse/Models/Options/CoursesOptions.cs
@@ -9,9 +9,10 @@ namespace MyCourse.Models.Options
 
     //contiene le opzioni relative al corso (paging, opzioni per ordinare i campi)
 
+    //classe autogenerata derivante dall'appsettings.json (Paste json as code)
     public partial class CoursesOptions
     {
-        public long PerPage { get; set; }
+        public int PerPage { get; set; }
 
         public CoursesOrderOptions Order { get; set; }
     }
